Change: long and lat in DB are now set to DECIMAL.

+++ db/schema/03_coords.sql
@@ -3,7 +3,7 @@ CREATE TABLE coords (
   id PRIMARY SERIAL KEY NOT NULL,
   map_id INTEGER REFERENCES maps(id) ON DELETE CASCADE,
   user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
-  longitude INTEGER,
-  latitude INTEGER,
+  longitude DECIMAL,
+  latitude DECIMAL,
   description text
 );
